refactor update password section to PasswordUpdateBox component
