+++ client/src/pages/Account/AccountSettings.jsx
@@ -1,11 +1,7 @@
-import { Box, TextField, Button } from '@mui/material';
-import { AccountDetailsBox } from './';
+import { Box, Button } from '@mui/material';
+import { AccountDetailsBox, PasswordUpdateBox } from './';
 
 function AccountSettings({ userName, userEmail }) {
-  const handlePasswordUpdate = () => {
-    return;
-  };
-
   const handleAccountDeletion = () => {
     return;
   };
@@ -13,36 +9,7 @@ function AccountSettings({ userName, userEmail }) {
   return (
     <div className="md:w-3/4 mx-auto">
       <AccountDetailsBox userName={userName} userEmail={userEmail} />
-      <Box className="bg-white p-4 rounded-md mt-4">
-        <h3 className="text-xl font-bold pb-2">Update Password</h3>
-        <TextField
-          id="password"
-          name="password"
-          label="Current Password"
-          defaultValue=""
-          margin="dense"
-          fullWidth
-        />
-        <TextField
-          id="newPassword"
-          name="newPassword"
-          label="New Password"
-          defaultValue=""
-          margin="dense"
-          fullWidth
-        />
-        <TextField
-          id="confirmNewPassword"
-          name="confirmNewPassword"
-          label="Confirm New Password"
-          defaultValue=""
-          margin="dense"
-          fullWidth
-        />
-        <Button onClick={handlePasswordUpdate} color="primary" sx={{ mt: 1 }}>
-          Update Password
-        </Button>
-      </Box>
+      <PasswordUpdateBox userName={userName} userEmail={userEmail} />
       <Box className="bg-white p-4 rounded-md mt-4">
         <h3 className="text-xl font-bold pb-2">Danger Zone</h3>
         <Button onClick={handleAccountDeletion} color="primary" sx={{ mt: 1 }}>
