+++ client/src/pages/Account/index.js
@@ -1,2 +1,3 @@
 export { default as AccountSettings } from './AccountSettings';
 export { default as AccountDetailsBox } from './AccountDetailsBox';
+export { default as PasswordUpdateBox } from './PasswordUpdateBox';
